refactor: 출력 파싱 기능 분리
